requirements.txt
torch
torchvision
wandb
PyYAML
matplotlib